+++ src/Three.WebGPU.js
@@ -7,8 +7,8 @@ export { default as BundleGroup } from './renderers/common/BundleGroup.js';
 export { default as QuadMesh } from './renderers/common/QuadMesh.js';
 export { default as PMREMGenerator } from './renderers/common/extras/PMREMGenerator.js';
 export { default as PostProcessing } from './renderers/common/PostProcessing.js';
-import * as PostProcessingUtils from './renderers/common/PostProcessingUtils.js';
-export { PostProcessingUtils };
+import * as RendererUtils from './renderers/common/RendererUtils.js';
+export { RendererUtils };
 export { default as StorageTexture } from './renderers/common/StorageTexture.js';
 export { default as StorageBufferAttribute } from './renderers/common/StorageBufferAttribute.js';
 export { default as StorageInstancedBufferAttribute } from './renderers/common/StorageInstancedBufferAttribute.js';
