+++ src/nodes/lighting/ShadowNode.js
@@ -16,7 +16,7 @@ import { renderGroup } from '../core/UniformGroupNode.js';
 import { viewZToLogarithmicDepth } from '../display/ViewportDepthNode.js';
 import { objectPosition } from '../accessors/Object3DNode.js';
 import { lightShadowMatrix } from '../accessors/Lights.js';
-import { resetRendererAndSceneState, restoreRendererAndSceneState } from '../../renderers/common/PostProcessingUtils.js';
+import { resetRendererAndSceneState, restoreRendererAndSceneState } from '../../renderers/common/RendererUtils.js';
 
 /** @module ShadowNode **/
