+++ src/renderers/common/RendererUtils.js
